# contracts/openapi/dashboard-v1.json
{
  "components": {
    "schemas": {
...
      "ConnectionFailureCategory": {
        "enum": [
          "invalid_credentials",
          "blocked_destination",
          "dns_failure",
          "tls_failure",
          "timeout",
          "rate_limited",
          "service_unavailable",
          "response_too_large",
          "incompatible_response",
          "credential_unavailable"
        ],
        "title": "ConnectionFailureCategory",
        "type": "string"
      },
      "ConnectionListResponse": {
        "additionalProperties": false,
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionSummary"
            },
            "title": "Connections",
            "type": "array"
          }
        },
        "required": [
          "connections"
        ],
        "title": "ConnectionListResponse",
        "type": "object"
      },
      "ConnectionStatus": {
        "enum": [
          "never_validated",
          "healthy",
          "degraded",
          "disabled",
          "deleted"
        ],
        "title": "ConnectionStatus",
        "type": "string"
      },
      "ConnectionSummary": {
        "additionalProperties": false,
        "properties": {
          "connection_type": {
            "$ref": "#/components/schemas/ConnectionType"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "last_failure_category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectionFailureCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_validation_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Validation Attempt At"
          },
          "last_validation_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Validation Success At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "normalized_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Base Url"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "usable": {
            "title": "Usable",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "connection_type",
          "status",
          "enabled",
          "is_default",
          "usable",
          "last_validation_attempt_at",
          "last_validation_success_at",
          "last_failure_category"
        ],
        "title": "ConnectionSummary",
        "type": "object"
      },
      "ConnectionType": {
        "enum": [
          "cloud",
          "remote"
        ],
        "title": "ConnectionType",
        "type": "string"
      },
      "CreateConnectionRequest": {
        "additionalProperties": false,
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "connection_type": {
            "$ref": "#/components/schemas/ConnectionType"
          },
          "credential": {
            "title": "Credential",
            "type": "string",
            "writeOnly": true
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "connection_type",
          "credential"
        ],
        "title": "CreateConnectionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplaceCredentialRequest": {
        "additionalProperties": false,
        "properties": {
          "credential": {
            "title": "Credential",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "credential"
        ],
        "title": "ReplaceCredentialRequest",
        "type": "object"
      },
...
      "UpdateConnectionRequest": {
        "additionalProperties": false,
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "UpdateConnectionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Parserium Collector",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/firecrawl/connections": {
      "get": {
        "operationId": "list_firecrawl_connections",
        "parameters": [
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Connections",
        "tags": [
          "firecrawl-connections"
        ]
      },
      "post": {
        "operationId": "create_firecrawl_connection",
        "parameters": [
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Connection",
        "tags": [
          "firecrawl-connections"
        ]
      }
    },
    "/api/v1/firecrawl/connections/{connection_id}": {
      "delete": {
        "operationId": "delete_firecrawl_connection",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Connection",
        "tags": [
          "firecrawl-connections"
        ]
      },
      "patch": {
        "operationId": "update_firecrawl_connection",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Connection",
        "tags": [
          "firecrawl-connections"
        ]
      }
    },
    "/api/v1/firecrawl/connections/{connection_id}/credential": {
      "put": {
        "operationId": "replace_firecrawl_connection_credential",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Credential",
        "tags": [
          "firecrawl-connections"
        ]
      }
    },
    "/api/v1/firecrawl/connections/{connection_id}/test": {
      "post": {
        "operationId": "test_firecrawl_connection",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection",
        "tags": [
          "firecrawl-connections"
        ]
      }
    },
...
  }
}